feat(eip8130): widen scope to u16 and match finalized ActorConfig/AccountState layout
Bring the native EIP-8130 types to storage/type parity with the finalized
Keystore contract:

- Scope widens u8 -> u16 across the wire type (Scope), InitialActor, the
  Eip8130Constants bits, ResolvedActor, the per-operation scope gate, and the
  TxAuthError::Scope field. RLP stays wire-compatible for values <= 0xff
  (minimal big-endian).
- ActorConfig is relaid out to authenticator(20) || expiry(6) || scope(2) ||
  reserved(4), matching the contract's field reorder and uint16 scope. The
  authorize ABI struct and the ActorAuthorized event actorData follow the same
  order (scope now 2 bytes).
- AccountState splits the local counter into localSequence(u32) || localEpoch(u32),
  widens lockUnion to uint48 (UNLOCKS_AT_MAX -> type(uint48).max) and
  defaultEOAScope to uint16, and relayouts the slot accordingly. is_initialized
  now also treats a non-zero local epoch as initialized.
- The CREATE2 actors commitment packs scope as 2 big-endian bytes, per the
  finalized _computeActorsCommitment.

Packing round-trip tests (to_word/from_word) and independent Solidity-style
packers are updated for every new offset, plus new coverage for the local-epoch
split and full-width uint48 lock union. No behavioral change to authorization
semantics; this is layout/width parity only.

+++ crates/execution/eip8130/src/tx_error.rs
@@ -31,7 +31,7 @@ pub enum TxAuthError {
         /// The operation whose scope requirement was not met.
         operation: Operation,
         /// The resolved actor's scope bitfield.
-        scope: u8,
+        scope: u16,
     },
 
     /// A config change or delegation targets a locked account. Both operations
